+++ src/Model/Entity/Photo.php
@@ -53,7 +53,7 @@ class Photo extends Entity
      * Virtual fields that should be exposed
      * @var array
      */
-    protected $_virtual = ['path', 'preview', 'thumbnail'];
+    protected $_virtual = ['path', 'preview'];
 
     /**
      * Gets the photo path (virtual field)
@@ -81,29 +81,11 @@ class Photo extends Entity
             return;
         }
 
-        $thumb = (new ThumbCreator($preview))->resize(1200)->save(['format' => 'jpg']);
+        $thumb = (new ThumbCreator($preview))->resize(1200, 1200)->save(['format' => 'jpg']);
         $preview = thumbUrl($thumb, true);
 
         list($width, $height) = getimagesize($thumb);
 
         return compact('preview', 'width', 'height');
-    }
-
-    /**
-     * Gets the photo thumbnail (virtual field)
-     * @return string|void Thumbnail path
-     * @uses _getPath()
-     */
-    protected function _getThumbnail()
-    {
-        $preview = $this->_getPath();
-
-        if (!$preview) {
-            return;
-        }
-
-        $thumb = (new ThumbCreator($preview))->resize(1200)->save(['format' => 'jpg']);
-
-        return thumbUrl($thumb, true);
     }
 }