+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # 2.x branch
 ## 2.17 branch
 ### 2.17.0
+* fixed bug when using multiple widgets that have the same name;
 * added `og:image:width` and `og:image:height` meta tags for previews;
 * added `preview` field to pages and posts tables, which contains url, width
     and height of the preview, encoded with json;
-* added `preview` and `thumbnail` virtual fields to photos;
+* added `preview` virtual field to photos;
 * improved `KcFinderComponent`. Now the class has `_getDefaultConfig()` and
     `initialize()` method;
 * improved `AuthComponent`. Now the class has the `initialize()` method;